Push new corrections

+++ src/files/pipe.c
@@ -51,7 +51,7 @@ static int pipe_resize_unlock_(pipe_t *pipe, size_t size)
         memcpy(remap, pipe->base, pipe->size);
         if (pipe->rpen < pipe->wpen) {
 
-        } else if (pipe->wpen - pipe->base < size - pipe->size) {
+        } else if ((size_t)(pipe->wpen - pipe->base) < size - pipe->size) {
             memcpy(remap + pipe->size, pipe->base, pipe->wpen - pipe->base);
             pipe->wpen += pipe->size;
         } else {
@@ -81,7 +81,7 @@ static int pipe_erase_unlock_(pipe_t *pipe, size_t len)
 {
     int bytes = 0;
     while (len > 0) {
-        size_t cap = MIN3(len, pipe->end - pipe->rpen, pipe->avail);
+        size_t cap = MIN3(len, (size_t)(pipe->end - pipe->rpen), pipe->avail);
         if (cap == 0)
             break;
         pipe->rpen += cap;
@@ -157,7 +157,7 @@ int pipe_write(pipe_t *pipe, const char *buf, size_t len, int flags)
             async_wait(&pipe->lock, &pipe->wlist, -1);
     }
     while (len > 0) {
-        int cap = MIN3(len, pipe->end - pipe->wpen, pipe->size - pipe->avail);
+        int cap = MIN3(len, (size_t)(pipe->end - pipe->wpen), pipe->size - pipe->avail);
         if (cap == 0) {
             if (pipe->size < pipe->max_size) {
 
@@ -206,7 +206,7 @@ int pipe_read(pipe_t *pipe, char *buf, size_t len, int flags)
             async_wait(&pipe->lock, &pipe->rlist, -1);
     }
     while (len > 0) {
-        int cap = MIN3(len, pipe->end - pipe->rpen, pipe->avail);
+        int cap = MIN3(len, (size_t)(pipe->end - pipe->rpen), pipe->avail);
         if (cap == 0) {
             async_raise(&pipe->wlist, 0);
             if (flags & IO_NO_BLOCK)
